fix: merge과정에서 발생한 에러 처리
+ properties에 username, password 값이 변경되어 있어 scoula, 1234로 수정했습니다.

+ RootConfig MapperScan에 report.mapper이 빠져있어 추가했습니다.

+ MemberExceptionHandler의 33번줄 handleDuplicateNickname 메소드의 어노테이션이 DuplicateEmailException으로 잘못 설정되어 있어 DuplicateNicknameException으로 수정하였습니다.

+++ backend/src/main/java/org/scoula/settlement/service/SettlementService.java
@@ -11,10 +11,10 @@ import java.util.List;
 public interface SettlementService {
   
     /* 정산 1단계에 필요한 요약 정보를 조회함. */
-    SettlementDTO.SettlementSummaryResponseDto getSettlementSummary(Long tripId);
+    SettlementDTO.SettlementSummaryResponseDto getSettlementSummary(int tripId);
 
     /* 정산 2단계: 최종 정산 결과를 계산하여 반환함. */
-    SettlementDTO.SettlementResultResponseDto calculateFinalSettlement(Long tripId);
+    SettlementDTO.SettlementResultResponseDto calculateFinalSettlement(int tripId);
 
     // ==================== 조회 관련 ====================
